+++ src/app/services/search_service.cpp
@@ -14,9 +14,9 @@
 #include <yams/daemon/resource/plugin_host.h>
 #endif
 #include <yams/plugins/search_provider_v1.h>
+#include <yams/search/parallel_post_processor.hpp>
 #include <yams/search/query_concept_extractor.h>
 #include <yams/search/query_qualifiers.hpp>
-#include <yams/search/parallel_post_processor.hpp>
 #include <yams/search/symbol_enrichment.h>
 
 #include <algorithm>